+++ Common/Services/Integrations/FauxCore/Enums/IconStyle.cs
@@ -1,0 +1,11 @@
+namespace StardewMods.Common.Services.Integrations.FauxCore;
+
+/// <summary>The styles of icon.</summary>
+public enum IconStyle
+{
+    /// <summary>An icon with a transparent background.</summary>
+    Transparent,
+
+    /// <summary>An icon with a button background.</summary>
+    Button,
+}