+++ FauxCore/Framework/Models/Icon.cs
@@ -8,20 +8,19 @@ using StardewValley.Menus;
 /// <inheritdoc />
 public sealed class Icon : IIcon
 {
-    private readonly Func<IIcon, ComponentStyle, ClickableTextureComponent> getComponent;
-    private readonly Func<IIcon, Texture2D> getTexture;
+    private readonly Func<IIcon, IconStyle, ClickableTextureComponent> getComponent;
+    private readonly Func<IIcon, IconStyle, Texture2D> getTexture;
 
     /// <summary>Initializes a new instance of the <see cref="Icon" /> class.</summary>
     /// <param name="getTexture">A function that returns the button texture.</param>
     /// <param name="getComponent">A function that return a new button.</param>
-    public Icon(Func<IIcon, Texture2D> getTexture, Func<IIcon, ComponentStyle, ClickableTextureComponent> getComponent)
+    public Icon(
+        Func<IIcon, IconStyle, Texture2D> getTexture,
+        Func<IIcon, IconStyle, ClickableTextureComponent> getComponent)
     {
         this.getTexture = getTexture;
         this.getComponent = getComponent;
     }
-
-    /// <inheritdoc />
-    public Texture2D Texture => this.getTexture(this);
 
     /// <inheritdoc />
     public Rectangle Area { get; set; } = Rectangle.Empty;
@@ -33,5 +32,8 @@ public sealed class Icon : IIcon
     public string Path { get; set; } = string.Empty;
 
     /// <inheritdoc />
-    public ClickableTextureComponent GetComponent(ComponentStyle style) => this.getComponent(this, style);
+    public ClickableTextureComponent GetComponent(IconStyle style) => this.getComponent(this, style);
+
+    /// <inheritdoc />
+    public Texture2D GetTexture(IconStyle style) => this.getTexture(this, style);
 }