+++ FauxCore/Framework/Models/ManagedTexture.cs
@@ -7,7 +7,6 @@ using StardewMods.Common.Services.Integrations.FauxCore;
 /// <inheritdoc />
 internal sealed class ManagedTexture : IManagedTexture
 {
-    private readonly IRawTextureData data;
     private readonly IGameContentHelper gameContentHelper;
 
     private Texture2D? cachedTexture;
@@ -18,29 +17,32 @@ internal sealed class ManagedTexture : IManagedTexture
     /// <param name="data">The raw data for the source texture.</param>
     public ManagedTexture(IGameContentHelper gameContentHelper, string path, IRawTextureData data)
     {
-        this.data = data;
+        this.RawData = data.Data;
+        this.Data = new Color[data.Data.Length];
+        Array.Copy(data.Data, this.Data, data.Data.Length);
+        this.Height = data.Height;
+        this.Width = data.Width;
         this.gameContentHelper = gameContentHelper;
         this.Name = gameContentHelper.ParseAssetName(path);
-        this.RawData = [..data.Data];
     }
 
     /// <inheritdoc />
-    public Color[] Data => this.data.Data;
+    public Color[] Data { get; }
 
     /// <inheritdoc />
-    public int Height => this.data.Height;
+    public int Height { get; }
 
     /// <inheritdoc />
     public IAssetName Name { get; }
 
-    /// <inheritdoc />
+    /// <summary>Gets the raw data of the texture.</summary>
     public Color[] RawData { get; }
 
     /// <inheritdoc />
     public Texture2D Value => this.cachedTexture ??= this.gameContentHelper.Load<Texture2D>(this.Name);
 
     /// <inheritdoc />
-    public int Width => this.data.Width;
+    public int Width { get; }
 
     /// <summary>Invalidates the texture cache.</summary>
     public void InvalidateCache() => this.cachedTexture = null;